+++ meetup2apricot/command_line.py
@@ -23,7 +23,7 @@ parser.add_argument(
     "--meetup-ids",
     dest="show_meetup_ids",
     action="store_true",
-    help="Show Meetup event IDs in reports.",
+    help="Show Meetup event IDs in reports",
 )
 
 parser.add_argument(
@@ -43,11 +43,12 @@ parser.add_argument(
 parser.add_argument(
     "-s",
     "--skip",
-    action="append",
+    action="extend",
+    nargs="+",
     default=[],
     dest="skip_meetup_ids",
     metavar="MEETUP_ID",
-    help="Skip the Meetup ID",
+    help="Skip events with these Meetup IDs",
 )
 
 parser.add_argument(