+++ HISTORY.rst
@@ -12,8 +12,10 @@ Future History
 Recent History
 ~~~~~~~~~~~~~~
 
-.. Next Release
-.. ------------------
+Next Release
+------------------
+
+* Allow multiple Meetup event IDs with a --skip command line option.
 
 1.6.0 (2021-03-24)
 ------------------